fixing issues with drive letter changes

+++ prepare_vm_disks.ps1
@@ -22,17 +22,20 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-$disks = Get-Disk | Where partitionstyle -eq 'raw' | sort number
-
-$letters = 70..89 | ForEach-Object { [char]$_ }
+$disks = Get-Disk | Sort Number
+# 83..89 = S..Y 
+$letters = 83..89 | ForEach-Object { [char]$_ } 
 $count = 0
 $label = "datadisk"
-
-foreach ($disk in $disks) {
-    $driveLetter = $letters[$count].ToString()
-    $disk | 
-    Initialize-Disk -PartitionStyle MBR -PassThru |
-    New-Partition -UseMaximumSize -DriveLetter $driveLetter |
-    Format-Volume -FileSystem NTFS -NewFileSystemLabel "$label.$count" -Confirm:$false -Force
-$count++
+ 
+for($index = 2; $index -lt $disks.Count; $index++) {
+    $driveLetter = $letters[$count].ToString()
+    if ($disks[$index].partitionstyle -eq 'raw') {
+        $disks[$index] | Initialize-Disk -PartitionStyle MBR -PassThru | 
+            New-Partition -UseMaximumSize -DriveLetter $driveLetter | 
+            Format-Volume -FileSystem NTFS -NewFileSystemLabel "$label.$count" -Confirm:$false -Force
+    } else {
+        $disks[$index] | Get-Partition | Set-Partition -NewDriveLetter $driveLetter
+    }
+    $count++
 }
